+++ src/apps/server/app.module.ts
@@ -7,9 +7,9 @@ import { APP_FILTER, APP_INTERCEPTOR, APP_PIPE } from '@nestjs/core';
 import { DatabaseModule } from 'src/modules/database/database.module';
 import { EnvModule } from 'src/modules/env/env.module';
 import { LogModule } from 'src/modules/log/log.module';
-import { ValidationException } from './exceptions/validation.exception';
-import { CustomExceptionFilter } from './filters/custom-exception.filter';
-import { LogInterceptor } from './interceptors/log.interceptor';
+import { ValidationException } from './common/exceptions/validation.exception';
+import { CustomExceptionFilter } from './common/filters/custom-exception.filter';
+import { LogInterceptor } from './common/interceptors/log.interceptor';
 import { SlackModule } from 'src/modules/slack/slack.module';
 import { AuthModule } from './auth/auth.module';
 import { AppController } from './app.controller';
