+++ src/apps/server/auth/auth.module.ts
@@ -1,17 +1,17 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { SigninGuard } from '../guards/signin.guard';
+import { SigninGuard } from '../common/guards/signin.guard';
 import { RedisCacheModule } from '../../../modules/cache/redis/redis.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { UserRepository } from '../../../modules/database/repositories/user.repository';
-import { JwtStrategy } from '../guards/strategies/jwt.strategy';
+import { JwtStrategy } from '../common/guards/strategies/jwt.strategy';
 import { UserInfoRepository } from '../../../modules/database/repositories/user-info.repository';
-import { JwtRefreshStrategy } from '../guards/strategies/jwt-refresh.strategy';
+import { JwtRefreshStrategy } from '../common/guards/strategies/jwt-refresh.strategy';
 import { ApiModule } from '../../../modules/api/api.module';
 import { APP_GUARD } from '@nestjs/core';
-import { JwtAuthGuard } from '../guards/jwt-auth.guard';
+import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
 
 @Module({
   imports: [
