+++ src/apps/server/auth/auth.service.ts
@@ -10,8 +10,8 @@ import { ConfigService } from '@nestjs/config';
 import {
   ACCESS_TOKEN_EXPIRES_IN,
   REFRESH_TOKEN_EXPIRES_IN,
-} from '../consts/jwt.const';
-import { UserPayload } from '../guards/signin-request-body.interface';
+} from '../common/consts/jwt.const';
+import { UserPayload } from '../common/guards/signin-request-body.interface';
 import { UserRepository } from '../../../modules/database/repositories/user.repository';
 import { CookieOptions } from 'express';
 import { UserInfoRepository } from '../../../modules/database/repositories/user-info.repository';
