+++ src/apps/server/common/exceptions/base.exception.ts
