+++ src/apps/server/common/exceptions/exception.type.ts
