+++ src/apps/server/common/exceptions/unknown.exception.ts
