+++ src/apps/server/common/exceptions/validation.exception.ts
