+++ src/apps/server/common/guards/strategies/jwt.strategy.ts
@@ -2,8 +2,8 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { UserRepository } from '../../../../modules/database/repositories/user.repository';
-import { UserJwtToken } from '../../auth/types/jwt-tokwn.type';
+import { UserRepository } from '../../../../../modules/database/repositories/user.repository';
+import { UserJwtToken } from '../../../auth/types/jwt-tokwn.type';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
