+++ src/apps/server/test/test.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { ACCESS_TOKEN_EXPIRES_IN } from '../consts/jwt.const';
+import { ACCESS_TOKEN_EXPIRES_IN } from '../common/consts/jwt.const';
 import { ConfigService } from '@nestjs/config';
 
 @Injectable()
