+++ src/enums/method.enum.ts
@@ -1,0 +1,7 @@
+export enum Method {
+  GET = 'GET',
+  POST = 'POST',
+  PUT = 'PuT',
+  DELETE = 'DELETE',
+  PATCH = 'PATCH',
+}
